requirements.txt
asgiref==3.11.0
asttokens==3.0.1
attrs==25.4.0
black==25.12.0
click==8.3.1
colorama==0.4.6
decorator==5.2.1
dj-database-url==3.0.1
Django==6.0
django-cors-headers==4.9.0
django-debug-toolbar==6.1.0
django-extensions==4.1
django-filter==25.2
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-spectacular==0.29.0
executing==2.2.1
flake8==7.3.0
inflection==0.5.1
ipython==9.8.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
matplotlib-inline==0.2.1
mccabe==0.7.0
mypy_extensions==1.1.0
packaging==25.0
parso==0.8.5
pathspec==0.12.1
platformdirs==4.5.1
prompt_toolkit==3.0.52
psycopg2==2.9.11
psycopg2-binary==2.9.11
pure_eval==0.2.3
pycodestyle==2.14.0
pyflakes==3.4.0
Pygments==2.19.2
PyJWT==2.10.1
python-decouple==3.8
python-dotenv==1.2.1
pytokens==0.3.0
PyYAML==6.0.3
referencing==0.37.0
rpds-py==0.30.0
sqlparse==0.5.5
stack-data==0.6.3
traitlets==5.14.3
tzdata==2025.3
uritemplate==4.2.0
wcwidth==0.2.14
